Mask category
Created new category for Mask and Inpaint
