+++ easycv/transforms/__init__.py
@@ -32,7 +32,8 @@ from easycv.transforms.spatial import (
     Translate,
     Paste,
 )
-from easycv.transforms.selectors import Select, Mask, Inpaint
+from easycv.transforms.mask import Mask, Inpaint
+from easycv.transforms.selectors import Select
 from easycv.transforms.detect import Scan, Eyes, Faces, Smile, Lines, Circles, Detect
 from easycv.transforms.draw import Draw
 from easycv.transforms.morphological import Erode, Dilate, Morphology
